added authentication to BlogIndView

+++ blogapp/views.py
@@ -4,7 +4,8 @@ from rest_framework.views import APIView
 from rest_framework import status
 from rest_framework.response import Response
 from rest_framework.viewsets import ModelViewSet
-
+from rest_framework.authentication import SessionAuthentication, BasicAuthentication
+from rest_framework.permissions import IsAdminUser, IsAuthenticated
 from userapp.models import Author, User
 from userapp.serializers import AuthorSerializer
 from .models import Blog, Tag
@@ -67,34 +68,53 @@ class BlogIndView(APIView):
     '''
     View to GET/PUT/DELETE individual blog-posts:
     '''
+    authentication_classes = [SessionAuthentication, BasicAuthentication]
+    permission_classes = [IsAuthenticated]
 
     def get(self, request, id: int):
-        queryset = Blog.objects.get(pk=id)
-        serialized = BlogPostSerializer(queryset)
+        if request.user or not request.user:
+            try:
+                queryset = Blog.objects.get(pk=id)
+            except Blog.DoesNotExist:
+                return Response(
+                {
+                    "error": f"Blog with ID #{id} does not exist."
+                },
+                status=status.HTTP_404_NOT_FOUND
+            )
+            serialized = BlogPostSerializer(queryset)
 
-        return Response(
-            serialized.data,
-            status=status.HTTP_302_FOUND
-        )
+            return Response(
+                serialized.data,
+                status=status.HTTP_302_FOUND
+            )
 
     def put(self, request, id: int):
         blog_post = Blog.objects.get(pk=id)
-        update = request.data
+        if request.user == blog_post.user or request.user.is_superuser:
+            update = request.data
 
-        updated = BlogPostSerializer(blog_post, data=update)
+            updated = BlogPostSerializer(blog_post, data=update)
 
-        if updated.is_valid():
-            updated.save()
-            return Response(
-                updated.data,
-                status=status.HTTP_202_ACCEPTED
-            )
-        else:
+            if updated.is_valid():
+                updated.save()
+                return Response(
+                    updated.data,
+                    status=status.HTTP_202_ACCEPTED
+                )
+            else:
+                return Response(
+                    {
+                        "error": str(updated.errors)
+                    },
+                    status=status.HTTP_400_BAD_REQUEST
+                )
+        elif request.user != blog_post.user or not request.user.is_superuser:
             return Response(
                 {
-                    "error": str(updated.errors)
+                    "error": f"you do not have permission to execute this ({__name__}) action."
                 },
-                status=status.HTTP_400_BAD_REQUEST
+                status=status.HTTP_401_UNAUTHORIZED
             )
 
     def delete(self, request, id: int):
@@ -102,11 +122,19 @@ class BlogIndView(APIView):
             blog_post = Blog.objects.get(pk=id)
             serialized = BlogPostSerializer(blog_post)
 
-            blog_post.delete()
+            if request.user == blog_post.user or request.user.is_superuser:
+                blog_post.delete()
 
-            return Response(
-                serialized.data,
-                status=status.HTTP_410_GONE
+                return Response(
+                    serialized.data,
+                    status=status.HTTP_410_GONE
+                )
+            elif request.user != blog_post.user or not request.user.is_superuser:
+                return Response(
+                {
+                    "error": f"you do not have permission to execute this ({__name__}) action."
+                },
+                status=status.HTTP_401_UNAUTHORIZED
             )
         except Blog.DoesNotExist:
             return Response(
